Add ContainerCats + Cats components
