+++ package.json
@@ -11,6 +11,7 @@
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-scripts": "3.0.1",
+    "styled-components": "^4.2.1",
     "typescript": "3.5.1"
   },
   "scripts": {
